use uint32 to avoid bigint error in js

+++ main.cpp
@@ -6,7 +6,7 @@
 
 extern "C" {
     EMSCRIPTEN_KEEPALIVE
-    float* decode_wav(uint8_t* data, size_t length, uint32_t* out_sample_rate, uint64_t* out_frame_count) {
+    float* decode_wav(uint8_t* data, size_t length, uint32_t* out_sample_rate, uint32_t* out_frame_count) {
         unsigned int channels;
         unsigned int sampleRate;
         drwav_uint64 frameCount;
@@ -18,9 +18,11 @@ extern "C" {
         if (samples == nullptr) return nullptr;
 
         *out_sample_rate = sampleRate;
-        *out_frame_count = frameCount;
+        *out_frame_count = (uint32_t)frameCount;
         return samples; // caller must free via free_wav()
     }
+
+
 
     EMSCRIPTEN_KEEPALIVE
     void free_wav(float* ptr) {
